Move GA events to /utils
