+++ src/pages/about/About.tsx
@@ -2,8 +2,9 @@ import React, { useEffect, useRef, useState } from 'react';
 import './About.less';
 import classnames from 'classnames';
 import { ABOUT_PARTNERS, LOCATION_INFO } from './About.config';
-import { usePageVisible, GAevent } from '../app/App.utils';
+import { usePageVisible } from '../app/App.utils';
 import { PageType } from '../app/App.config';
+import { GAevent } from '../../utils';
 import gsap from 'gsap';
 
 export const About: React.FC = () => {
