+++ src/pages/app/App.utils.tsx
@@ -10,7 +10,6 @@ import React, {
 import { PageType } from './App.config';
 import EE from 'eventemitter3';
 import { isFunction, mapValues, mean, pick } from 'lodash-es';
-import ReactGA from 'react-ga4';
 
 
 interface AppContextValue {
@@ -133,56 +132,3 @@ LoadingSourceTypeArray.forEach((type) => {
         }
     });
 });
-
-
-/** GA **/
-
-const GAtiming = {
-    startTime : -1,
-    endTime : -1,
-    prevView: '',
-    calcTiming: function(webview: string) {
-        let diff = 0;
-        if (this.startTime === -1) { // begin timing
-            this.startTime = new Date().getTime(); 
-        }
-        else if (this.prevView !== webview) { // changed webview
-            this.endTime = new Date().getTime(); 
-
-            diff = this.endTime -  this.startTime;
-            GAevent("timing", this.prevView.split('-')[0] + '-timing', diff);
-            
-            this.startTime = this.endTime;
-        }
-        this.prevView = webview;
-    }
-}
-
-const GA_ID : string = (process.env.REACT_APP_GOOGLE_ANALYTICS_ID as string);
-
-export const initGA = () => {
-    // initialize ga
-    ReactGA.initialize(GA_ID);
-}
-
-export const GAevent = (category: string, action:string, value=-1) => {
-    // event tracking
-    if (value === -1) {
-        ReactGA.event({
-            category: category,
-            action: action
-        });
-    }
-    else {
-        ReactGA.event({
-            category: category,
-            action: action,
-            value: value
-        });
-    }
-
-    if (category === "webview") {
-        GAtiming.calcTiming(action);
-    }
-    
-}
