+++ src/pages/poster/Poster.tsx
@@ -6,11 +6,11 @@ import 'swiper/css';
 import 'swiper/css/autoplay';
 import 'swiper/css/effect-fade';
 import './Poster.less';
-import { loadingEE, LoadingSourceType, usePageVisible, GAevent } from '../app/App.utils';
+import { loadingEE, LoadingSourceType, usePageVisible } from '../app/App.utils';
 import { PageType } from '../app/App.config';
 import gsap from 'gsap';
 import { POSTER_FEATURES } from './Poster.config';
-import { IS_MOBILE } from '../../utils';
+import { IS_MOBILE, GAevent } from '../../utils';
 Swiper.use([Autoplay, EffectFade]);
 
 export const Poster: React.FC = () => {
