+++ src/pages/wall/Wall.tsx
@@ -2,7 +2,8 @@ import gsap from 'gsap';
 import React, { useEffect, useRef, useState } from 'react';
 import { ButterflyGL } from '../../components/butterfly-gl/ButterflyGL';
 import { PageType } from '../app/App.config';
-import { usePageVisible, GAevent } from '../app/App.utils';
+import { usePageVisible } from '../app/App.utils';
+import { GAevent } from '../../utils';
 import './Wall.less';
 
 export const Wall: React.FC = () => {
